Better MemSQL node support
Summary:
We now need a flag to know if the current node is a MemSQL node to save
us a fairly expensive checking query every 60 seconds.  The alternative
would be to not require the memsql flag, but run the find node routine
on every node every 60 seconds.  I am not sure if this is reasonable
since this system needs to safely scale to thousands of nodes.

Test Plan: eyes

Reviewers: ankur, nikita

Differential Revision: http://grizzly.memsql.com/D3498