Better master node detection
Summary:
Often the Dashboard uses 127.0.0.1 in its config for the master node.
This indirectly causes the row referencing the master node in the nodes
table to have the host "127.0.0.1".  This would later indirectly cause
the collectd plugin to in-correctly or not assign a node_alias to the
master node.

To fix the issue I use master_connection on the random_aggregator_pool to get the actual master's host, check it against my own ip's and then update the node_alias table with the master node_id.

Test Plan: none

Reviewers: chase

Differential Revision: http://grizzly.memsql.com/D4018